pyproject.toml
[tool.poetry]
name = "word2pdf"
version = "0.1.0"
description = "Cli to convert Word files to PDF files."
authors = ["hakshu <flyhighup.air25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
docx2pdf = "^0.1.8"
python-fire = "^0.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"
mypy = "^0.982"
pyproject-flake8 = "^5.0.4.post1"
flake8-isort = "^5.0.0"
flake8-quotes = "^3.3.1"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py310"]
include = '\.py$'

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 79
inline-quotes = '"'
exclude = [
  ".git",
  ".venv",
  "*_cache"
]

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = [
  "docx2pdf",
  "fire"
]
ignore_missing_imports = true
